satellite-application/src/main/kotlin/kr/hs/entrydsm/satellite/domain/document/usecase/SubmitMyDocumentUseCase.kt
package kr.hs.entrydsm.satellite.domain.document.usecase

import kr.hs.entrydsm.satellite.common.annotation.UseCase
import kr.hs.entrydsm.satellite.domain.auth.spi.SecurityPort
import kr.hs.entrydsm.satellite.domain.document.domain.DocumentStatus
import kr.hs.entrydsm.satellite.domain.document.exception.DocumentIllegalStatusException
import kr.hs.entrydsm.satellite.domain.document.exception.DocumentNotFoundException
import kr.hs.entrydsm.satellite.domain.document.spi.DocumentPort

@UseCase
class SubmitMyDocumentUseCase(
    private val securityPort: SecurityPort,
    private val documentPort: DocumentPort
) {
    suspend fun execute() {

        val student = securityPort.getCurrentStudent()
        val document = documentPort.queryByWriterStudentId(student.id) ?: throw DocumentNotFoundException

        if (document.status != DocumentStatus.CREATED) {
            throw DocumentIllegalStatusException
        }

        documentPort.save(
            document.apply { changeStatus(DocumentStatus.SUBMITTED) }
        )
    }
}